fixed conventions and folders, also fixed by comments
