+++ my-app/src/Components/CommentComponents/Comment/Comment.css
@@ -1,4 +1,4 @@
-.Comment{
+.comment{
   width: fit-content;
   height: fit-content;
   padding: 1vw;
