+++ my-app/src/Components/CommentComponents/CommentList/CommentList.css
@@ -1,4 +1,4 @@
-.CommentList{
+.comment-list{
     width: fit-content;
     height: fit-content;
     padding: 1vw;
