+++ my-app/src/Components/Post/Post.css
@@ -1,4 +1,4 @@
-.Post{
+.post{
     width: fit-content;
   height: fit-content;
   padding: 1vw;
