+++ my-app/src/Components/Post/Post.tsx
@@ -1,4 +1,4 @@
-import React, { ReactElement, } from "react";
+import React, { ReactElement, ReactNode, } from "react";
 import Like from "../Like/Like";
 import CommentList from "../CommentComponents/CommentList/CommentList";
 import PostModel from "../../Models/PostModel/PostModel";
@@ -7,10 +7,11 @@ import './Post.css'
 
 interface Props{
     postModel : PostModel
+    children? : ReactNode
 }
 
 const Post = ({postModel} : Props):ReactElement => {
-    return <div className="Post">
+    return <div className="post">
         <h2>{postModel.publisherName}</h2>
         <p>{postModel.content}</p>
         <p>{`${postModel.publishDate.toDateString()} ${postModel.publishDate.toLocaleTimeString()}`}</p>
